Blueprint class and model maker, with future structure. Next step : blueprint, filter and routes generation
