+++ Hook/cmd.py
@@ -1,7 +1,3 @@
-from flask.ext.mongoengine import MongoEngine 
-from flask.ext.github import GitHub
-from flask.ext.login import LoginManager
-
 import yaml
 import os
 
@@ -13,9 +9,5 @@ def read_yaml(path):
 def write_env(conf, app):
     for key, value in conf.items():
         app.config[key] = value
-
-db = MongoEngine()
-github_api = GitHub()
-login_manager = LoginManager()
 
 conf, env = read_yaml(os.path.join(os.path.dirname(__file__), '../config.yaml'))
