+++ Hook/ext.py
@@ -1,0 +1,1176 @@
+from flask import Flask, Blueprint, url_for, request, render_template, g, Markup, session, redirect, jsonify
+from flask.ext.mongoengine import MongoEngine
+from flask.ext.github import GitHub
+from flask.ext.login import LoginManager
+
+from pkg_resources import resource_filename
+import re
+import math
+import hmac
+import hashlib
+import json
+import requests
+from uuid import uuid4
+
+from Hook.models import model_maker
+from Hook.common import slugify
+
+class HookUI(object):
+    ROUTES = [
+        ('/', "r_index", ["GET"]),
+        ('/login/form', "r_login", ["GET"]),
+        ('/logout', "r_logout", ["GET"]),
+
+        ('/api/github/callback', "r_github_oauth", ["GET"]),
+        ('/api/github/payload', "r_github_payload", ["GET"]),
+        ("/api/hooktest", "api_hooktest_endpoint", ["POST"]),
+
+        ('/repo/<owner>/<repository>', "r_repository", ["GET", "POST"]),
+        ('/repo/<owner>/<repository>/<uuid>', "r_repository_test", ["GET"]),
+
+        ("/api/rest/v1.0/user/repositories", "r_api_user_repositories", ["GET", "POST"]),
+        ("/api/rest/v1.0/user/repositories/<owner>/<name>", "r_api_user_repository_switch", ["PUT"]),
+
+        ('/api/rest/v1.0/code/<owner>/<repository>/status.svg'"r_repo_badge_status", ["GET"]),
+        ('/api/rest/v1.0/code/<owner>/<repository>/cts.svg'"r_repo_cts_status", ["GET"]),
+        ('/api/rest/v1.0/code/<owner>/<repository>/coverage.svg'"r_repo_badge_coverage", ["GET"]),
+        ('/api/rest/v1.0/code/<owner>/<repository>/test'"r_api_test_generate_route", ["GET"]),
+        ('/api/rest/v1.0/code/<owner>/<repository>', "r_api_repo_history", ["GET", "DELETE"]),
+        ('/api/rest/v1.0/code/<owner>/<repository>/unit', "r_api_repo_unit_history", ["GET"])
+    ]
+
+    FILTERS = [
+        "r_nice_ref",
+        "r_nice_branch",
+        'r_slugify',
+        'r_checked',
+        'r_btn',
+        'r_format_log',
+        'r_tei',
+        'r_epidoc',
+        'r_success_class',
+        "r_ctsized"
+    ]
+
+    VERBOSE = re.compile("(>>>>>>[^>]+)")
+    PR_FINDER = re.compile("pull\/([0-9]+)\/head")
+
+    def __init__(self,
+         mongo=None, github=None, login=None,
+         remote=None, github_secret=None, hooktest_secret=None,
+         static_folder=None, template_folder=None, app=None, name=None
+    ):
+        """
+
+        :param mongo:
+        :param github:
+        :param login:
+        :param static_folder:
+        :param template_folder:
+        :param app:
+        :param name:
+        :return:
+        """
+        self.app = app
+        self.name = name
+        self.blueprint = None
+
+        self.db = mongo
+        self.api = github
+        self.login_manager = login
+
+        self.remote = remote
+        self.github_secret = github_secret
+        self.hooktest_secret = hooktest_secret
+
+        self.static_folder = static_folder
+        if not self.static_folder:
+            self.static_folder = resource_filename("Hook", "data/static")
+
+        self.template_folder = template_folder
+        if not self.template_folder:
+            self.template_folder = resource_filename("Hook", "data/templates")
+
+        if self.name is None:
+            self.name = __name__
+
+        if self.app:
+            self.init_app(app=app)
+
+    def before_request(self):
+        """ Before request function for the Blueprint
+        """
+        if 'user_id' in self.session:
+            self.g.user = self.m_User.objects.get(uuid=self.session['user_id'])
+
+    def init_app(self, app):
+        if not self.db:
+            self.db = MongoEngine(app)
+        if not self.api:
+            self.api = GitHub(app)
+        if not self.login_manager:
+            self.login_manager = LoginManager(app)
+
+        if not self.remote:
+          self.remote = app.config["HOOKUI_REMOTE"]
+        if not self.github_secret:
+          self.github_secret = app.config["HOOKUI_GITHUB_SECRET"]
+        if not self.hooktest_secret:
+          self.hooktest_secret = app.configp["HOOKUI_HOOKTEST_SECRET"]
+
+        # Register token getter for github
+        self.api.access_token_getter(self.github_token_getter)
+
+        # Register the user loader for LoginManager
+        self.login_manager.user_loader(self.login_manager_user_loader)
+
+        # Register the before request
+        self.blueprint.before_request(self.before_request)
+
+        # Generate Instance models
+        self.m_User, self.m_Repository, self.m_RepoTest, self.m_DocLogs, self.m_DocTest, self.m_DocUnitStatus = model_maker(self.db)
+
+    def login_manager_user_loader(self, user_id):
+        """ Load a user
+
+        :param user_id: User id
+        :return:
+        """
+        if hasattr(self.g, "user"):
+            return self.g.user
+        return None
+
+    def github_token_getter(self):
+        """ Get the github token
+
+        :return: User github access token
+        """
+        if hasattr(self.g, "user"):
+            if self.g.user is not None:
+                return self.g.user.github_access_token
+
+    def r_login(self):
+        """ Route for login
+        """
+        with self.app.app_context():
+            return self.login(url_for(".index"))
+
+    def r_logout(self):
+        """ Route for logout
+        """
+        with self.app.app_context():
+            return self.logout(url_for(".index"))
+
+    def r_github_oauth(self, access_token):
+        """ GitHub oauth route
+        """
+        def func(self, access_token):
+            with self.app.app_context():
+                next_uri = request.args.get('next') or url_for('.index')
+                return self.authorize(access_token, request, success=next_uri, error=url_for(".index"))
+
+        self.api.authorized_handler(func)(access_token)
+
+    def r_github_payload(self):
+        """ GitHub payload route
+        """
+        return self.handle_payload(
+            request,
+            request.headers,
+            callback_url=url_for(".api_hooktest_endpoint", _external=True)
+        )
+
+    def r_index(self):
+        """ Index route
+        """
+        return render_template("index.html")
+
+    def r_repository(self, owner, repository):
+        """ Route for the repository test history, statistics and settings
+
+        :param owner: Name of the owner
+        :param repository: Name of the repository
+        """
+        kwargs = self.read_repo(owner, repository, request)
+        return render_template(
+            'repo.html',
+            **kwargs
+        )
+
+    def r_repository_test(self, owner, repository, uuid):
+        """ Route for a test report
+
+        :param owner: Name of the owner
+        :param repository: Name of the repository
+        :param uuid: UUID of the test
+        """
+        kwargs, status, header = self.repo_report(owner, repository, uuid)
+        if status == 200:
+            return render_template("report.html", **kwargs)
+        else:
+            return kwargs, status, header
+
+    def r_api_hooktest_endpoint(self):
+        """ Route HookTest endpoint
+        """
+        return "", self.handle_hooktest_log(request), {}
+
+    def r_api_user_repositories(self):
+        """ Route fetching user repositories
+
+        """
+        return self.fetch(request.method)
+
+    def r_api_user_repository_switch(self, owner, repository):
+        """ Route for toggling github webhook
+
+        :param owner: Name of the user
+        :param repository: Name of the repository
+        """
+        with self.app.app_context():
+            return self.link(owner, repository, url_for(".api_test_payload"))
+
+    def r_repo_badge_status(self, owner, repository):
+        """ Get a Badge for a repo
+
+        :param owner: Name of the user
+        :param repository: Name of the repository
+        """
+        response, status, header = self.status_badge(owner, repository, branch=request.args.get("branch"), uuid=request.args.get("uuid"))
+        if response:
+            return render_template(response), status, header
+        else:
+            return "", status, {}
+
+    def r_repo_cts_status(self, owner, repository):
+        """ Get a Badge for a repo
+
+        :param owner: Name of the user
+        :param repository: Name of the repository
+        """
+        response, kwargs, status, header = self.cts_badge(owner, repository, branch=request.args.get("branch"), uuid=request.args.get("uuid"))
+        if response:
+            return render_template(response, **kwargs), status, header
+        else:
+            return "", status, {}
+
+    def r_repo_badge_coverage(self, owner, repository):
+        """ Get a Badge for a repo
+
+        :param owner: Name of the user
+        :param repository: Name of the repository
+        """
+        response, kwargs, status, header = self.coverage_badge(owner, repository, branch=request.args.get("branch"), uuid=request.args.get("uuid"))
+        if response:
+            return render_template(response, **kwargs), status, header
+        else:
+            return "", status, {}
+
+    def r_api_test_generate_route(self, owner, repository):
+        """ Generate a test on the machine
+
+        :param owner: Name of the user
+        :param repository: Name of the repository
+        """
+        with self.app.app_context():
+            return self.generate(
+                owner,
+                repository,
+                callback_url=url_for(".api_hooktest_endpoint", _external=True),
+                check_user=True
+            )
+
+    def r_api_repo_history(self, owner, repository):
+        """ Return json history of previous tests
+
+        :param owner: Name of the user
+        :param repository: Name of the repository
+        """
+        if request.method == "DELETE":
+            return self.cancel(owner, repository, uuid=request.args.get("uuid"))
+        elif request.args.get("uuid"):
+            return self.repo_report(
+                owner,
+                repository,
+                uuid=request.args.get("uuid"),
+                start=request.args.get("start", 0, type=int),
+                limit=request.args.get("limit", None, type=int),
+                json=True
+            )
+        else:
+            return self.history(owner, repository)
+
+    def r_api_repo_unit_history(self, owner, repository):
+        """ Return json representation of one unit test
+
+        :param owner: Name of the user
+        :param repository: Name of the repository
+        """
+        return self.repo_report_unit(
+            owner,
+            repository,
+            uuid=request.args.get("uuid"),
+            unit=request.args.get("unit", "all")
+
+        )
+
+    @staticmethod
+    def f_nice_ref(branch, commit):
+        """ Transform a git formatted reference into something more human readable or returns part of commit sha
+
+        :param branch: Github Reference
+        :param commit: Commit SHA
+        :return: Human readable reference
+        """
+        if HookUI.PR_FINDER.match(branch):
+            return "PR #{0}".format(branch.strip("pull/").strip("/head"))
+        return commit[0:8]
+
+    @staticmethod
+    def f_nice_branch(branch):
+        """ Transform a git formatted reference into something more human readable
+
+        :param branch: Github Reference
+        :return: Human readable branch name
+        """
+        if HookUI.PR_FINDER.match(branch):
+            return "PR #{0}".format(branch.strip("pull/").strip("/head"))
+        else:
+            return branch.split("/")[-1]
+
+    @staticmethod
+    def f_slugify(string):
+        """ Slugify filter
+
+        :return: Slugified string
+        """
+        if not string:
+            return ""
+        return slugify(string)
+
+    @staticmethod
+    def f_checked_bool(boolean):
+        """ Check a checkbox if boolean is true
+
+        :param boolean: Boolean
+        :return: "checked" if boolean is True
+        """
+        if boolean:
+            return " checked "
+        return ""
+
+    @staticmethod
+    def f_btn_bool(boolean):
+        """ Return btn bs3 class depending on status
+
+        :param boolean: Boolean
+        :return: "btn-success" and "btn-danger" depending on boolean
+        """
+        if boolean:
+            return "btn-success"
+        return "btn-danger"
+
+    @staticmethod
+    def f_format_log(string):
+        """ Format log string output from HookTest
+
+        :param string: Log
+        :return: Formatted log
+        """
+        if not string:
+            return ""
+        else:
+            if string.startswith(">>> "):
+                string = Markup("<u>{0}</u>".format(string.strip(">>> ")))
+            elif string.startswith(">>>> "):
+                string = Markup("<b>{0}</b>".format(string.strip(">>>> ")))
+            elif string.startswith(">>>>> "):
+                string = Markup("<i>{0}</i>".format(string.strip(">>>>> ")))
+            elif HookUI.VERBOSE.findall(string):
+                string = Markup("</li><li>".join(["<span class='verbose'>{0}</span>".format(found.strip(">>>>>> ")) for found in HookUI.VERBOSE.findall(string)]))
+            elif string.startswith("[success]"):
+                string = Markup("<span class='success'>{0}</span>".format(string.strip("[success]")))
+            elif string.startswith("[failure]"):
+                string = Markup("<span class='failure'>{0}</span>".format(string.strip("[failure]")))
+            return string
+
+    @staticmethod
+    def f_check_tei(string):
+        """ Check if value is "tei"
+
+        :param string: String to check against "tei"
+        :return: Checked class
+        """
+        if string == "tei":
+            return "checked"
+        return ""
+
+    @staticmethod
+    def f_check_epidoc(string):
+        """ Check if value is "epidoc"
+
+        :param string: String to check against "epidoc"
+        :return: Checked class
+        """
+        if string == "epidoc":
+            return "checked"
+        return ""
+
+    @staticmethod
+    def f_success_class(status):
+        """ Return success or failed depending on boolean
+
+        :param status: Status
+        :return: Success or fail
+        """
+        string = ""
+        if status is True:
+            string = "success"
+        elif status is False:
+            string = "failed"
+        return string
+
+    @staticmethod
+    def f_ctsized(cts_tuple):
+        """ Join a tuple of text
+
+        :param cts_tuple: Tuple representing the number of texts cts compliant and the total number of text
+        :type cts_tuple: (int, int)
+        :return: Stringified tuple
+        """
+        return "{0}/{1}".format(*cts_tuple)
+
+    """
+        CONTROLLER FUNCTIONS
+    """
+
+    def read_repo(self, owner, repository, request):
+        """ Read the repository tests
+
+        :param owner:
+        :param repository:
+        :param request:
+        :return:
+        .. todo:: Pagination
+        """
+
+        start, end = 0, 20
+        repository = self.m_Repository.objects.get_or_404(owner__iexact=owner, name__iexact=repository)
+
+        if request.method == "POST" and hasattr(self.g, "user") and self.g.user in repository.authors:
+            repository.config(request.form)
+
+        # PAGINATION !!!
+        tests = self.m_RepoTest.objects(
+            repository=repository
+        ).exclude(
+            "units"
+        )
+
+        done = [test for test in tests if test.finished]
+        running = [test for test in tests if test.finished == False]
+
+        for r in running:
+            if r.total > 0:
+                r.percent = int(r.tested / r.total * 100)
+            else:
+                r.percent = 0
+
+        return {
+            "repository": repository,
+            "tests": done,
+            "running": running
+        }
+
+    def repo_report(self, owner, repository, uuid, start=0, limit=None, json=False):
+        """ Generate data for repository report
+
+        :param owner:
+        :param repository:
+        :param uuid:
+        :param start:
+        :param limit:
+        :param json:
+        :return:
+        """
+
+        repository = self.m_Repository.objects.get_or_404(owner__iexact=owner, name__iexact=repository)
+        test = self.m_RepoTest\
+            .objects(repository=repository, uuid=uuid)\
+            .exclude("units.text_logs")\
+            .first()
+
+        if json is True:
+            return jsonify(units=test.units_status())
+        else:
+            return {
+                "repository": repository,
+                "test": test
+            }, 200, {}
+
+    def repo_report_unit(self, owner, repository, uuid, unit):
+        """ Generate data for repository report for a single unit
+
+        :param owner:
+        :param repository:
+        :param uuid:
+        :param unit:
+        :return:
+        """
+
+        repository = self.m_Repository.objects.get_or_404(owner__iexact=owner, name__iexact=repository)
+        if unit == "all":
+            test = self.m_RepoTest.objects.get_or_404(repository=repository, uuid=uuid)
+            report = self.m_RepoTest.report(owner, repository, repo_test=test)
+        else:
+            test = self.m_RepoTest.objects(
+                repository=repository, uuid=uuid
+            ).first().units.filter(
+                path=unit
+            ).first()
+            if test is None:
+                return "", 404
+            report = self.m_DocTest.report(test)
+
+        return jsonify(report)
+
+    def generate_informations(self, repository):
+        """ Generate informations for a user generated build
+
+        :param repository: Repository for which a test should be run
+        :type repository: Repository
+        :return:
+        """
+        status = self.api.get(
+            "repos/{owner}/{name}/commits".format(owner=repository.owner, name=repository.name),
+            params={"sha": "master", "per_page": "1"}
+        )
+        if len(status) == 0:
+            return "error", "No commits available", 404
+
+        sha = status[0]["sha"]
+        ref = "refs/heads/master"
+        if "author" in status[0]:
+            creator = status[0]["author"]["login"]
+        else:
+            creator = status[0]["commit"]["author"]["name"]
+        guid = str(uuid4())
+        url = status[0]["html_url"]
+
+        return ref, creator, sha, url, guid
+
+    def generate(self, username, reponame, callback_url=None, ref=None, creator=None, sha=None, url=None, uuid=None, check_branch=False, check_user=True):
+        """ Generate a test on the machine
+
+        :param username: Name of the user
+        :type username: str
+        :param reponame: Name of the repository
+        :type reponame: str
+        :param callback_url: URL to send log to
+        :param ref: branch to be tested
+        :type ref: str
+        :param creator: Person responsible for starting the test
+        :param sha: SHA of the commit
+        :param url: URL of the resource
+        :param uuid: UUID to use
+        :param check_branch: If set to True, should check against repo.master_pr
+        """
+        repo = self.m_Repository.objects.get_or_404(owner__iexact=username, name__iexact=reponame)
+
+        if check_user:
+            if hasattr(self.g, "user") and not repo.isWritable(self.g.user):
+                resp = jsonify(status="error", message="You are not an owner of the repository", uuid=None)
+                return resp
+
+        if check_branch == True and repo.master_pr == True:
+            if not HookUI.PR_FINDER.match(ref) and not ref.endswith("master"):
+                response = jsonify(status="ignore", message="Test ignored because this is not a pull request nor a push to master")
+                return response
+
+        if creator is None:  # sha and url should be None
+            informations = self.generate_informations(repo)
+            if len(informations) == 3:
+                return informations
+            else:
+                ref, creator, sha, url, uuid = informations
+
+        avatar = "https://avatars.githubusercontent.com/{0}".format(creator)
+
+        test = self.m_RepoTest.Get_or_Create(
+            uuid=uuid,
+            repository=repo,
+            branch=ref,
+            user=creator,
+            gravatar=avatar,
+            sha=sha,
+            link=url
+        )
+
+        status, message = self.dispatch(test, callback_url)
+        self.comment(test)
+
+        return jsonify(status=status, message=message, uuid=uuid)
+
+    def dispatch(self, test, callback_url):
+        """ Dispatch a test to the redis queue
+
+        :param test: The test to be sent
+        :param callback_url: URL to send log to
+        """
+
+        params = {
+            "repository": test.repository.full_name,
+            "ping": callback_url,
+            "verbose": test.verbose,
+            "console": False,
+            "uuid" : test.uuid,
+            "scheme": test.scheme,
+            "branch": test.branch
+        }
+        params = bytes(json.dumps(params), encoding="utf-8")
+        response = requests.put(self.remote, data=params, headers={'content-type': 'application/json', "HookTest-Secure-X" : self.make_hooktest_signature(params)})
+        infos = response.json()
+
+        if infos["status"] == "queud":
+            test.update(hash=infos["job_id"], status="queued")
+            return "success", "Test launched"
+        else:
+            return infos["status"], "Error while dispatching test"
+
+    def handle_hooktest_log(self, request):
+        """ Handle data received from Redis server commands
+
+        :param request: request object
+        :return: Status Code
+        """
+
+        if not request.data or not request.headers.get("HookTest-UUID"):
+            return 404
+        elif self.check_hooktest_signature(request.data, request.headers.get("HookTest-Secure-X")) is False:
+            return 401
+
+        # Now we get the repo
+        uuid = request.headers.get("HookTest-UUID")
+        test = self.m_RepoTest.objects(uuid=uuid).exclude("units.logs").exclude("units.text_logs").first()
+        if not test:
+            return 404
+        data = json.loads(request.data.decode('utf-8'))
+
+        # If the test just started
+        if "status" in data:
+            test.update(status=data["status"])
+        if "message" in data:
+            test.update(error_message=data["message"])
+
+        if "files" in data and "inventories" in data and "texts" in data:
+            test.update(
+                texts=data["texts"],
+                cts_metadata=data["inventories"]
+            )
+        else:
+            # If we have units, we have single logs to save
+            update = {}
+            if "units" in data:
+                _units = []
+                for unit in data["units"]:
+                    _units.append(self.m_DocTest(
+                        at=unit["at"],
+                        path=unit["name"],
+                        status=unit["status"],
+                        coverage=unit["coverage"]
+                    ))
+
+                    for text in unit["logs"]:
+                        _units[-1].text_logs.append(self.m_DocLogs(text=text))
+
+                    for test_name, test_status in unit["units"].items():
+                        _units[-1].logs.append(self.m_DocUnitStatus(
+                            title=test_name,
+                            status=test_status
+                        ))
+                update["push_all__units"] = _units
+
+            if "coverage" in data:
+                update["coverage"] = data["coverage"]
+                update["status"] = data["status"]
+
+            if len(update) > 0:
+                test.update(**update)
+
+        if "status" in data and data["status"] in ["success", "error", "failed"]:
+            test.reload()
+            self.comment(test)
+        return 200
+
+    def make_hooktest_signature(self, body):
+        """ Check the signature sent by a request with the body
+
+        :param body: Raw body of the request
+        :return:
+        """
+        return '{0}'.format(
+            hmac.new(
+                bytes(self.hooktest_secret, encoding="utf-8"),
+                body,
+                hashlib.sha1
+            ).hexdigest()
+        )
+
+    def check_hooktest_signature(self, body, hook_signature):
+        """ Check the signature sent by a request with the body
+
+        :param body: Raw body of the request
+        :param hook_signature: Signature sent by github
+        :return:
+        """
+        if self.make_hooktest_signature(body) == hook_signature:
+            return True
+        else:
+            return False
+
+    def check_github_signature(self, body, hub_signature):
+        """ Check the signature sent by a request with the body
+
+        :param body: Raw body of the request
+        :param hub_signature: Signature sent by github
+        :return:
+        """
+        signature = 'sha1={0}'.format(hmac.new(bytes(self.github_secret, encoding="utf-8"), body, hashlib.sha1).hexdigest())
+        if signature == hub_signature:
+            return True
+        else:
+            return False
+
+    def handle_payload(self, request, headers, callback_url):
+        """ Handle a payload call from Github
+
+        :param request:
+        :param headers:
+        :return:
+        """
+        status, message, code = "error", "Webhook query is not handled", 200
+        creator, sha, ref, url, do = None, None, None, None, None
+        pull_request = False
+
+        signature = headers.get("X-Hub-Signature")
+        if not self.check_github_signature(request.data, signature):
+            response = jsonify(
+                status="error",
+                message="Signature check did not pass"
+            )
+            response.status_code = 300
+            return response
+
+        payload = request.get_json(force=True)
+        #  guid = headers.get("X-GitHub-Delivery")
+        event = headers.get("X-GitHub-Event")
+        username, repository = tuple(payload["repository"]["full_name"].split("/"))
+        if event in ["push", "pull_request"]:
+            if event == "push":
+                creator = payload["head_commit"]["committer"]["username"]
+                sha = payload["head_commit"]["id"]
+                url = payload["compare"]
+                ref = payload["ref"]
+                pull_request = False
+                do = True
+            elif event == "pull_request" and payload["action"] in ["reopened", "opened", "synchronize"]:
+                creator = payload["pull_request"]["user"]["login"]
+                url = payload["pull_request"]["html_url"]
+                sha = payload["pull_request"]["head"]["sha"]
+                ref = "pull/{0}/head".format(payload["number"])
+                do = True
+                pull_request = True
+            if do:
+                response = self.generate(
+                    username,
+                    repository,
+                    callback_url=callback_url,
+                    creator=creator,
+                    sha=sha,
+                    url=url,
+                    ref=ref,
+                    uuid=str(uuid4()),
+                    check_branch=pull_request
+                )
+                return response
+
+        response = jsonify(status=status, message=message)
+        response.status_code = code
+        return response
+
+    def cancel(self, owner, repository, uuid=None):
+        """ Cancel a test
+
+        :param owner:
+        :param repository:
+        :param uuid:
+        :return:
+        """
+        if uuid is None:
+            return "Unknown test", 404, {}
+        repo = self.m_Repository.objects.get_or_404(owner__iexact=owner, name__iexact=repository)
+
+        if hasattr(self.g, "user") and not repo.isWritable(self.g.user):
+            resp = jsonify(cancelled=False, message="You are not an owner of the repository", uuid=None)
+            return resp
+
+        test = self.m_RepoTest.objects.get_or_404(repository=repo, uuid__iexact=uuid)
+        test.update(status="error")
+
+        response = requests.delete("{0}/{1}".format(self.remote, uuid))
+        return jsonify(cancelled=True)
+
+    def link(self, owner, repository, callback_url):
+        """ Set Github to start or stop pinging the current repository
+
+        :param owner:
+        :param repository:
+        :param callback_url: URL for callback
+        :return: json
+        """
+        response = False
+
+        if hasattr(self.g, "user") and self.g.user:
+            repository = self.m_Repository.objects(
+                authors__in=[self.g.user],
+                owner__iexact=owner,
+                name__iexact=repository
+            )
+            if len(repository) > 0:
+                repository = repository.first()
+                tested = not repository.tested
+                repository.update(tested=tested)
+                repository.reload()
+                response = self.__make_link(repository, callback_url)
+
+        return jsonify(status=response)
+
+    def __make_link(self, repository, callback_url):
+        """ Create or delete hooks on GitHub API
+
+        :param repository: Repository to add or delete the hook from
+        :type repository: Hook.models.github.Repository
+
+        :returns: Active status
+        """
+        uri = "repos/{owner}/{repo}/hooks".format(owner=repository.owner, repo=repository.name)
+        payload = self.domain + callback_url
+
+        if repository.tested is True:
+            # Create hooks
+            hook_data = {
+              "name": "web",
+              "active": True,
+              "events": [
+                "push",
+                "pull_request"
+              ],
+              "config": {
+                "url": payload,
+                "content_type": "json",
+                "secret": self.github_secret
+              }
+            }
+            service = self.api.post(uri, data=hook_data)
+            if "id" in service:
+                repository.update(hook_id=service["id"])
+        else:
+            if repository.hook_id is None:
+                hooks = self.api.get(uri)
+                hooks = [service for service in hooks if service["config"]["url"] == payload]
+                if len(hooks) == 0:
+                    uuid = None
+                else:
+                    uuid = hooks[0]["id"]
+            else:
+                uuid = repository.hook_id
+            if uuid is not None:
+                self.api.delete("repos/{owner}/{repo}/hooks/{id}".format(owner=repository.owner, repo=repository.name, id=uuid))
+
+        return repository.tested
+
+    def cts_badge(self, username, reponame, branch=None, uuid=None):
+        """
+
+        :param username:
+        :param reponame:
+        :param branch:
+        :param uuid:
+        :return:
+        :return: (Template, Kwargs, Status Code, Headers)
+        """
+        repo = self.repo(
+            owner=username,
+            name=reponame,
+            branch=branch,
+            uuid=uuid
+        )
+
+        if not repo:
+            return None, None, 404, {}
+
+        cts, total = repo.ctsized()
+        template = "svg/cts.xml"
+        return template, {"cts": cts, "total": total}, 200, {'Content-Type': 'image/svg+xml; charset=utf-8'}
+
+    def status_badge(self, username, reponame, branch=None, uuid=None):
+        """ Create a status badge
+
+        :param username:
+        :param reponame:
+        :param branch:
+        :param uuid:
+        :return: (Template, Headers, Status Code)
+        """
+        repo = self.repo(
+            owner=username,
+            name=reponame,
+            branch=branch,
+            uuid=uuid
+        )
+
+        if not repo:
+            return None, 404, {}
+        else:
+            template = "svg/build.{0}.xml".format(repo.status)
+
+        return template, 200, {'Content-Type': 'image/svg+xml; charset=utf-8'}
+
+    def coverage_badge(self, username, reponame, branch=None, uuid=None):
+        """ Create a status badge
+
+        :param username:
+        :param reponame:
+        :param branch:
+        :param uuid:
+        :return: (Template, Kwargs, Status Code, Headers)
+        """
+        repo = self.repo(
+            owner=username,
+            name=reponame,
+            branch=branch,
+            uuid=uuid
+        )
+
+        if not repo or not repo.coverage:
+            return "svg/build.coverage.unknown.xml", {}, 200, {'Content-Type': 'image/svg+xml; charset=utf-8'}
+
+        score = math.floor(repo.coverage * 100) / 100
+
+        if repo.coverage > 90:
+            template = "svg/build.coverage.success.xml"
+        elif repo.coverage > 75:
+            template = "svg/build.coverage.acceptable.xml"
+        else:
+            template = "svg/build.coverage.failure.xml"
+
+        return template, {"score": score}, 200, {'Content-Type': 'image/svg+xml; charset=utf-8'}
+
+    def repo(self, owner, name, branch=None, uuid=None):
+        """ Helper to find a repository
+
+        :param owner:
+        :param name:
+        :param branch:
+        :param uuid:
+        :return:
+        """
+        repo = self.m_Repository.objects.get_or_404(owner__iexact=owner, name__iexact=name)
+        test_args = {
+            "repository": repo
+        }
+        if branch:
+            test_args["branch__iexact"] = branch
+
+        if uuid:
+            test_args["uuid__iexact"] = uuid
+
+        test = self.m_RepoTest.objects(**test_args).exclude("units.logs").exclude("units.text_logs")
+
+        return test.first()
+
+    def history(self, username, reponame):
+        """ Return the history of a repo
+
+        :param username:
+        :param reponame:
+        :return:
+        """
+        repository = self.m_Repository.objects.get_or_404(owner__iexact=username, name__iexact=reponame)
+        history = {
+            "username": username,
+            "reponame": reponame,
+            "logs": [
+                {
+                    "run_at": event.run_at,
+                    "uuid": event.uuid,
+                    "coverage": event.coverage,
+                    "ref": event.branch,
+                    "slug": event.branch_slug
+                }
+                for event in self.m_RepoTest.objects.exclude("units")(repository=repository)
+            ]
+        }
+        return jsonify(**history)
+
+    def comment(self, test):
+        """ Takes care of sending information to github API through the comment / status API
+
+        :param test: The test currently running or finished
+        :return:
+        """
+        repo = test.repository
+        uri = "repos/{owner}/{repo}/statuses/{sha}".format(owner=repo.owner, repo=repo.name, sha=test.sha)
+        if test.status == "error":
+            state = "error"
+            sentence = "Test cancelled or ran into an error"
+        elif test.status == "success":
+            state = "success"
+            sentence = "Full repository is cts compliant"
+        elif test.status == "failed":
+            state = "failure"
+            sentence = "{0:.2f}% of unit tests passed".format(test.coverage)
+        else:
+            state = "pending"
+            sentence = "Currently testinself.g..."
+
+        data = {
+          "state": state,
+          "target_url": self.domain+"/repo/{username}/{reponame}/{uuid}".format(username=repo.owner, reponame=repo.name, uuid=test.uuid),
+          "description": sentence,
+          "context": "continuous-integration/capitains-hook"
+        }
+
+        params = {}
+        if not hasattr(self.g, "user"):
+            user = repo.authors[0]
+            access_token = user.github_access_token
+            params = {"access_token": access_token}
+
+        print(uri, data, params)
+        self.api.post(uri, data=data, params=params)
+
+
+    @staticmethod
+    def slice(elements, start, max=None):
+        """ Return a sublist of elements, starting from index {start} with a length of {max}
+
+        :param elements: List of elements to slice
+        :type elements: list
+        :param start: Starting element (0-based Index)
+        :type start: int
+        :param max: Maximum number of elements to return
+        :type max: int
+
+        :return: (Sublist, Starting index of the sublist, end_index)
+        :rtype: (list, int, int)
+        """
+        length = len(elements)
+        real_index = start + 1
+        if length < real_index:
+            return [], length, 0
+        else:
+            if max is not None and max + start < length:
+                return elements[start:max+start], start, max+start
+            else:
+                return elements[start:], start, length - 1
+
+    def login(self, url_redirect):
+        """ Login the user using github API
+
+        :param url_redirect: Url to redirect to
+        :return: redirect(url_redirect)
+        """
+        if self.session.get('user_id', None) is None:
+            return self.api.authorize(scope=",".join(["user:email", "repo:status", "admin:repo_hook", "read:org"]))
+        return redirect(url_redirect)
+
+
+    """
+        USER CONTROLLER
+    """
+    def authorize(self, access_token, request, success, error):
+        """ Callback uri of Github API to signin/up a user
+
+        :param access_token: Access token sent by github
+        :param request: request informations
+        :param success: success redirect uri
+        :param error: error redirect uri
+        :return: redirect()
+        """
+        if access_token is None:
+            return redirect(error)
+
+        user = self.m_User.objects(github_access_token=access_token)
+        if len(user) == 0:
+            # Make a call to the API
+            more = self.api.get("user", params={"access_token": access_token})
+            user = self.m_User(
+                uuid=str(uuid4()),
+                github_access_token=access_token,
+                mail=more["email"],
+                git_id=more["id"],
+                login=more["login"]
+            )
+            user.save()
+        else:
+            user = user.first()
+
+        self.session['user_id'] = user.uuid
+        return redirect(success)
+
+    def fetch(self, method):
+        """ Fetch repositories of a user
+
+        :param method:
+        :return:
+        """
+        response = []
+        if hasattr(self.g, "user") and self.g.user:
+            if method == "POST":
+                # We clear the old authors
+                self.g.user.remove_authorship()
+
+                repositories = self.api.get(
+                    "user/repos",
+                    params={
+                        "affiliation": "owner,collaborator,organization_member",
+                        "access_token": self.g.user.github_access_token
+                    },
+                    all_pages=True
+                )
+
+                for repo in repositories:
+                    owner = repo["owner"]["login"]
+                    name = repo["name"]
+                    repo = self.g.user.addto(owner=owner, name=name)
+                    response.append(repo)
+            elif method == "GET":
+                response = self.g.user.repositories
+
+        return jsonify({"repositories" : [repo.dict() for repo in response]})
+
+    def logout(self, url_redirect):
+        """ Logout the user and redirect to a specific url
+
+        :param url_redirect: Url to redirect to
+        :return: redirect(url_redirect)
+        """
+        self.session.pop('user_id', None)
+        return redirect(url_redirect)
+
+    @property
+    def session(self):
+        with self.app.app_context():
+            return session
+
+    @property
+    def g(self):
+        with self.app.app_context():
+            return g
+
+    @property
+    def domain(self):
+        with self.app.app_context():
+            return url_for("", _external=True)
+
